+++ src/main/java/com/sparta/deliverybackend/api/oauth2/config/InMemoryProviderRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.deliverybackend.api.oauth2.repository;
+package com.sparta.deliverybackend.api.oauth2.config;
 
 import java.util.HashMap;
 import java.util.Map;
