+++ src/main/java/com/sparta/deliverybackend/api/oauth2/service/Oauth2Service.java
@@ -20,10 +20,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sparta.deliverybackend.api.auth.controller.dto.LoginResDto;
-import com.sparta.deliverybackend.api.oauth2.attributes.Oauth2Attributes;
+import com.sparta.deliverybackend.api.oauth2.config.InMemoryProviderRepository;
 import com.sparta.deliverybackend.api.oauth2.controller.dto.OauthMemberProfile;
 import com.sparta.deliverybackend.api.oauth2.provider.Oauth2Provider;
-import com.sparta.deliverybackend.api.oauth2.repository.InMemoryProviderRepository;
 import com.sparta.deliverybackend.common.JwtHelper;
 import com.sparta.deliverybackend.domain.member.entity.Member;
 import com.sparta.deliverybackend.domain.member.repository.MemberRepository;
